Task to analysise partial accuracy for transactions
For a set of postcodes, that we know have matched LAs for real users,
fetch the LA matched for the full postcode, and repeat for the partial
(outcode) postcode, so we see how often this is likely to give an
different answers (as partial postcode location is based on a centroid,
and may match against a different LA vs the full. Because maps are fun!).

Fetches the artefact so we can trial this against different transactions
that may have different selection criteria for area types (see
`LocationIdentifier`).

A private controller method has the area selection logic, so re-use that
in a slightly hacky way, so that the selection process will be the same
as live.

This has a companion GDS API adaptors change which adds
`location_for_partial_postcode` and builds a response that looks almost
the same as the expected `Location` object, and can be operated on in
the same way by the controller methods, ensuring a fair comparison.